CameraX: Actually select the highest quality video by default
Prior to this commit SnapdragonCamera assumes that the HAL returns the
list of supported video sizes in descending order of quality and
simply picks the first one from the list as the default quality.

Instead, find the first matching entry of pref_video_quality_entryvalues
that the HAL indicated was supported.

Change-Id: Ifea79e0e16a9015557539e098317536a32b9ff1f
Signed-off-by: Alex Naidis <alex.naidis@linux.com>