CameraX: don't crash when pref is not boolean
In some situation, a portion of pref (enableRecordingLocation) is
created without first calling upgradeGlobalPreference first.
Leaving the pref without the version field and makes the application
thinks that the pref is in old version.

When it tries to do an upgrade on the pref later, it tries to getBoolean
for that location field, but that location field is already in string
format, making getBoolean throws an exception.

The best is to call upgradeGlobalPreference and upgradeLocalPreference
when the application starts.

However it should be no harm just add a simple guard in getBoolean
to avoid that sitation.

FEIJ-1258

Change-Id: Ic73078556b1a198a58968806091f8b0afd1ad6cc
Signed-off-by: Alex Naidis <alex.naidis@linux.com>